refactor error types and mod layout
