+++ benches/format_str.rs
@@ -3,7 +3,7 @@ extern crate test;
 
 use std::io::Write;
 use test::Bencher;
-use uuid::prelude::*;
+use uuid::Uuid;
 
 #[bench]
 fn bench_hyphen(b: &mut Bencher) {
