+++ benches/invalid_parse_str.rs
@@ -2,7 +2,7 @@
 extern crate test;
 
 use test::Bencher;
-use uuid::prelude::*;
+use uuid::Uuid;
 
 #[bench]
 fn bench_parse_invalid_strings(b: &mut Bencher) {
