+++ src/adapter/core_support.rs
