+++ src/lib.rs
@@ -114,7 +114,7 @@
 //!
 //! [`wasm-bindgen`]: https://github.com/rustwasm/wasm-bindgen
 
-#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(not(test), no_std)]
 #![deny(
     missing_copy_implementations,
     missing_debug_implementations,
@@ -126,22 +126,25 @@
     html_root_url = "https://docs.rs/uuid/0.7.4"
 )]
 
+#[cfg(feature = "std")]
+extern crate std;
+
+#[cfg(not(feature = "std"))]
+extern crate core as std;
+
+mod error;
+mod prelude;
+mod builder;
+mod parser;
+
 pub mod adapter;
-pub mod builder;
-pub mod parser;
-pub mod prelude;
 #[cfg(feature = "v1")]
 pub mod v1;
 
-pub use crate::builder::Builder;
-
-mod core_support;
 #[cfg(feature = "serde")]
 mod serde_support;
 #[cfg(feature = "slog")]
 mod slog_support;
-#[cfg(feature = "std")]
-mod std_support;
 #[cfg(test)]
 mod test_util;
 #[cfg(all(
@@ -180,49 +183,15 @@ mod v5;
 #[cfg(all(windows, feature = "winapi"))]
 mod winapi_support;
 
+use crate::std::{fmt, str};
+
+pub use crate::{
+    error::Error,
+    builder::Builder,
+};
+
 /// A 128-bit (16 byte) buffer containing the ID.
 pub type Bytes = [u8; 16];
-
-/// The error that can occur when creating a [`Uuid`].
-///
-/// [`Uuid`]: struct.Uuid.html
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct BytesError {
-    expected: usize,
-    found: usize,
-}
-
-/// A general error that can occur when handling [`Uuid`]s.
-///
-/// Although specialized error types exist in the crate,
-/// sometimes where particular error type occurred is hidden
-/// until errors need to be handled. This allows to enumerate
-/// the errors.
-///
-/// [`Uuid`]: struct.Uuid.html
-// TODO: improve the doc
-// BODY: This detail should be fine for initial merge
-
-// TODO: write tests for Error
-// BODY: not immediately blocking, but should be covered for 1.0
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub enum Error {
-    /// An error occurred while handling [`Uuid`] bytes.
-    ///
-    /// See [`BytesError`]
-    ///
-    /// [`BytesError`]: struct.BytesError.html
-    /// [`Uuid`]: struct.Uuid.html
-    Bytes(BytesError),
-
-    /// An error occurred while parsing a [`Uuid`] string.
-    ///
-    /// See [`parser::ParseError`]
-    ///
-    /// [`parser::ParseError`]: parser/enum.ParseError.html
-    /// [`Uuid`]: struct.Uuid.html
-    Parse(parser::ParseError),
-}
 
 /// The version of the UUID, denoting the generating algorithm.
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -260,28 +229,6 @@ pub enum Variant {
 #[derive(Clone, Copy, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Uuid(Bytes);
 
-impl BytesError {
-    /// The expected number of bytes.
-    #[inline]
-    pub const fn expected(&self) -> usize {
-        self.expected
-    }
-
-    /// The number of bytes found.
-    #[inline]
-    pub const fn found(&self) -> usize {
-        self.found
-    }
-
-    /// Create a new [`UuidError`].
-    ///
-    /// [`UuidError`]: struct.UuidError.html
-    #[inline]
-    pub const fn new(expected: usize, found: usize) -> Self {
-        BytesError { expected, found }
-    }
-}
-
 impl Uuid {
     /// [`Uuid`] namespace for Domain Name System (DNS).
     ///
@@ -314,259 +261,6 @@ impl Uuid {
         0x6b, 0xa7, 0xb8, 0x14, 0x9d, 0xad, 0x11, 0xd1, 0x80, 0xb4, 0x00, 0xc0,
         0x4f, 0xd4, 0x30, 0xc8,
     ]);
-
-    /// The 'nil UUID'.
-    ///
-    /// The nil UUID is special form of UUID that is specified to have all
-    /// 128 bits set to zero, as defined in [IETF RFC 4122 Section 4.1.7][RFC].
-    ///
-    /// [RFC]: https://tools.ietf.org/html/rfc4122.html#section-4.1.7
-    ///
-    /// # Examples
-    ///
-    /// Basic usage:
-    ///
-    /// ```
-    /// use uuid::Uuid;
-    ///
-    /// let uuid = Uuid::nil();
-    ///
-    /// assert_eq!(
-    ///     uuid.to_hyphenated().to_string(),
-    ///     "00000000-0000-0000-0000-000000000000"
-    /// );
-    /// ```
-    pub const fn nil() -> Self {
-        Uuid::from_bytes([0; 16])
-    }
-
-    /// Creates a `Uuid` from four field values in big-endian order.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if `d4`'s length is not 8 bytes.
-    ///
-    /// # Examples
-    ///
-    /// Basic usage:
-    ///
-    /// ```
-    /// use uuid::Uuid;
-    ///
-    /// let d4 = [12, 3, 9, 56, 54, 43, 8, 9];
-    ///
-    /// let uuid = Uuid::from_fields(42, 12, 5, &d4);
-    /// let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());
-    ///
-    /// let expected_uuid =
-    ///     Ok(String::from("0000002a-000c-0005-0c03-0938362b0809"));
-    ///
-    /// assert_eq!(expected_uuid, uuid);
-    /// ```
-    ///
-    /// An invalid length:
-    ///
-    /// ```
-    /// use uuid::prelude::*;
-    ///
-    /// let d4 = [12];
-    ///
-    /// let uuid = uuid::Uuid::from_fields(42, 12, 5, &d4);
-    ///
-    /// let expected_uuid = Err(uuid::BytesError::new(8, d4.len()));
-    ///
-    /// assert_eq!(expected_uuid, uuid);
-    /// ```
-    pub fn from_fields(
-        d1: u32,
-        d2: u16,
-        d3: u16,
-        d4: &[u8],
-    ) -> Result<Uuid, BytesError> {
-        const D4_LEN: usize = 8;
-
-        let len = d4.len();
-
-        if len != D4_LEN {
-            return Err(BytesError::new(D4_LEN, len));
-        }
-
-        Ok(Uuid::from_bytes([
-            (d1 >> 24) as u8,
-            (d1 >> 16) as u8,
-            (d1 >> 8) as u8,
-            d1 as u8,
-            (d2 >> 8) as u8,
-            d2 as u8,
-            (d3 >> 8) as u8,
-            d3 as u8,
-            d4[0],
-            d4[1],
-            d4[2],
-            d4[3],
-            d4[4],
-            d4[5],
-            d4[6],
-            d4[7],
-        ]))
-    }
-
-    /// Creates a `Uuid` from four field values in little-endian order.
-    ///
-    /// The bytes in the `d1`, `d2` and `d3` fields will
-    /// be converted into big-endian order.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use uuid::Uuid;
-    ///
-    /// let d1 = 0xAB3F1097u32;
-    /// let d2 = 0x501Eu16;
-    /// let d3 = 0xB736u16;
-    /// let d4 = [12, 3, 9, 56, 54, 43, 8, 9];
-    ///
-    /// let uuid = Uuid::from_fields_le(d1, d2, d3, &d4);
-    /// let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());
-    ///
-    /// let expected_uuid =
-    ///     Ok(String::from("97103fab-1e50-36b7-0c03-0938362b0809"));
-    ///
-    /// assert_eq!(expected_uuid, uuid);
-    /// ```
-    pub fn from_fields_le(
-        d1: u32,
-        d2: u16,
-        d3: u16,
-        d4: &[u8],
-    ) -> Result<Uuid, BytesError> {
-        const D4_LEN: usize = 8;
-
-        let len = d4.len();
-
-        if len != D4_LEN {
-            return Err(BytesError::new(D4_LEN, len));
-        }
-
-        Ok(Uuid::from_bytes([
-            d1 as u8,
-            (d1 >> 8) as u8,
-            (d1 >> 16) as u8,
-            (d1 >> 24) as u8,
-            (d2) as u8,
-            (d2 >> 8) as u8,
-            d3 as u8,
-            (d3 >> 8) as u8,
-            d4[0],
-            d4[1],
-            d4[2],
-            d4[3],
-            d4[4],
-            d4[5],
-            d4[6],
-            d4[7],
-        ]))
-    }
-
-    /// Creates a `Uuid` from a 128bit value in big-endian order.
-    pub const fn from_u128(v: u128) -> Self {
-        Uuid::from_bytes([
-            (v >> 120) as u8,
-            (v >> 112) as u8,
-            (v >> 104) as u8,
-            (v >> 96) as u8,
-            (v >> 88) as u8,
-            (v >> 80) as u8,
-            (v >> 72) as u8,
-            (v >> 64) as u8,
-            (v >> 56) as u8,
-            (v >> 48) as u8,
-            (v >> 40) as u8,
-            (v >> 32) as u8,
-            (v >> 24) as u8,
-            (v >> 16) as u8,
-            (v >> 8) as u8,
-            v as u8,
-        ])
-    }
-
-    /// Creates a `Uuid` from a 128bit value in little-endian order.
-    pub const fn from_u128_le(v: u128) -> Self {
-        Uuid::from_bytes([
-            v as u8,
-            (v >> 8) as u8,
-            (v >> 16) as u8,
-            (v >> 24) as u8,
-            (v >> 32) as u8,
-            (v >> 40) as u8,
-            (v >> 48) as u8,
-            (v >> 56) as u8,
-            (v >> 64) as u8,
-            (v >> 72) as u8,
-            (v >> 80) as u8,
-            (v >> 88) as u8,
-            (v >> 96) as u8,
-            (v >> 104) as u8,
-            (v >> 112) as u8,
-            (v >> 120) as u8,
-        ])
-    }
-
-    /// Creates a `Uuid` using the supplied big-endian bytes.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if `b` has any length other than 16.
-    ///
-    /// # Examples
-    ///
-    /// Basic usage:
-    ///
-    /// ```
-    /// use uuid::Uuid;
-    ///
-    /// let bytes = [4, 54, 67, 12, 43, 2, 98, 76, 32, 50, 87, 5, 1, 33, 43, 87];
-    ///
-    /// let uuid = Uuid::from_slice(&bytes);
-    /// let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());
-    ///
-    /// let expected_uuid =
-    ///     Ok(String::from("0436430c-2b02-624c-2032-570501212b57"));
-    ///
-    /// assert_eq!(expected_uuid, uuid);
-    /// ```
-    ///
-    /// An incorrect number of bytes:
-    ///
-    /// ```
-    /// use uuid::prelude::*;
-    ///
-    /// let bytes = [4, 54, 67, 12, 43, 2, 98, 76];
-    ///
-    /// let uuid = Uuid::from_slice(&bytes);
-    ///
-    /// let expected_uuid = Err(uuid::BytesError::new(16, 8));
-    ///
-    /// assert_eq!(expected_uuid, uuid);
-    /// ```
-    pub fn from_slice(b: &[u8]) -> Result<Uuid, BytesError> {
-        const BYTES_LEN: usize = 16;
-
-        let len = b.len();
-
-        if len != BYTES_LEN {
-            return Err(BytesError::new(BYTES_LEN, len));
-        }
-
-        let mut bytes: Bytes = [0; 16];
-        bytes.copy_from_slice(b);
-        Ok(Uuid::from_bytes(bytes))
-    }
-
-    /// Creates a `Uuid` using the supplied big-endian bytes.
-    pub const fn from_bytes(bytes: Bytes) -> Uuid {
-        Uuid(bytes)
-    }
 
     /// Returns the variant of the `Uuid` structure.
     ///
@@ -784,157 +478,6 @@ impl Uuid {
         Some((ts, count))
     }
 
-    /// Parses a `Uuid` from a string of hexadecimal digits with optional
-    /// hyphens.
-    ///
-    /// Any of the formats generated by this module (simple, hyphenated, urn)
-    /// are supported by this parsing function.
-    pub fn parse_str(mut input: &str) -> Result<Uuid, parser::ParseError> {
-        // Ensure length is valid for any of the supported formats
-        let len = input.len();
-
-        if len == adapter::Urn::LENGTH && input.starts_with("urn:uuid:") {
-            input = &input[9..];
-        } else if !parser::len_matches_any(
-            len,
-            &[adapter::Hyphenated::LENGTH, adapter::Simple::LENGTH],
-        ) {
-            return Err(parser::ParseError::InvalidLength {
-                expected: parser::ExpectedLength::Any(&[
-                    adapter::Hyphenated::LENGTH,
-                    adapter::Simple::LENGTH,
-                ]),
-                found: len,
-            });
-        }
-
-        // `digit` counts only hexadecimal digits, `i_char` counts all chars.
-        let mut digit = 0;
-        let mut group = 0;
-        let mut acc = 0;
-        let mut buffer = [0u8; 16];
-
-        for (i_char, chr) in input.bytes().enumerate() {
-            if digit as usize >= adapter::Simple::LENGTH && group != 4 {
-                if group == 0 {
-                    return Err(parser::ParseError::InvalidLength {
-                        expected: parser::ExpectedLength::Any(&[
-                            adapter::Hyphenated::LENGTH,
-                            adapter::Simple::LENGTH,
-                        ]),
-                        found: len,
-                    });
-                }
-
-                return Err(parser::ParseError::InvalidGroupCount {
-                    expected: parser::ExpectedLength::Any(&[1, 5]),
-                    found: group + 1,
-                });
-            }
-
-            if digit % 2 == 0 {
-                // First digit of the byte.
-                match chr {
-                    // Calulate upper half.
-                    b'0'..=b'9' => acc = chr - b'0',
-                    b'a'..=b'f' => acc = chr - b'a' + 10,
-                    b'A'..=b'F' => acc = chr - b'A' + 10,
-                    // Found a group delimiter
-                    b'-' => {
-                        // TODO: remove the u8 cast
-                        // BODY: this only needed until we switch to
-                        //       ParseError
-                        if parser::ACC_GROUP_LENS[group] as u8 != digit {
-                            // Calculate how many digits this group consists of
-                            // in the input.
-                            let found = if group > 0 {
-                                // TODO: remove the u8 cast
-                                // BODY: this only needed until we switch to
-                                //       ParseError
-                                digit - parser::ACC_GROUP_LENS[group - 1] as u8
-                            } else {
-                                digit
-                            };
-
-                            return Err(
-                                parser::ParseError::InvalidGroupLength {
-                                    expected: parser::ExpectedLength::Exact(
-                                        parser::GROUP_LENS[group],
-                                    ),
-                                    found: found as usize,
-                                    group,
-                                },
-                            );
-                        }
-                        // Next group, decrement digit, it is incremented again
-                        // at the bottom.
-                        group += 1;
-                        digit -= 1;
-                    }
-                    _ => {
-                        return Err(parser::ParseError::InvalidCharacter {
-                            expected: "0123456789abcdefABCDEF-",
-                            found: input[i_char..].chars().next().unwrap(),
-                            index: i_char,
-                            urn: parser::UrnPrefix::Optional,
-                        });
-                    }
-                }
-            } else {
-                // Second digit of the byte, shift the upper half.
-                acc *= 16;
-                match chr {
-                    b'0'..=b'9' => acc += chr - b'0',
-                    b'a'..=b'f' => acc += chr - b'a' + 10,
-                    b'A'..=b'F' => acc += chr - b'A' + 10,
-                    b'-' => {
-                        // The byte isn't complete yet.
-                        let found = if group > 0 {
-                            // TODO: remove the u8 cast
-                            // BODY: this only needed until we switch to
-                            //       ParseError
-                            digit - parser::ACC_GROUP_LENS[group - 1] as u8
-                        } else {
-                            digit
-                        };
-
-                        return Err(parser::ParseError::InvalidGroupLength {
-                            expected: parser::ExpectedLength::Exact(
-                                parser::GROUP_LENS[group],
-                            ),
-                            found: found as usize,
-                            group,
-                        });
-                    }
-                    _ => {
-                        return Err(parser::ParseError::InvalidCharacter {
-                            expected: "0123456789abcdefABCDEF-",
-                            found: input[i_char..].chars().next().unwrap(),
-                            index: i_char,
-                            urn: parser::UrnPrefix::Optional,
-                        });
-                    }
-                }
-                buffer[(digit / 2) as usize] = acc;
-            }
-            digit += 1;
-        }
-
-        // Now check the last group.
-        // TODO: remove the u8 cast
-        // BODY: this only needed until we switch to
-        //       ParseError
-        if parser::ACC_GROUP_LENS[4] as u8 != digit {
-            return Err(parser::ParseError::InvalidGroupLength {
-                expected: parser::ExpectedLength::Exact(parser::GROUP_LENS[4]),
-                found: (digit as usize - parser::ACC_GROUP_LENS[3]),
-                group,
-            });
-        }
-
-        Ok(Uuid::from_bytes(buffer))
-    }
-
     /// Tests if the UUID is nil
     pub fn is_nil(&self) -> bool {
         self.as_bytes().iter().all(|&b| b == 0)
@@ -971,13 +514,162 @@ impl Uuid {
     }
 }
 
+impl fmt::Debug for Uuid {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::LowerHex::fmt(self, f)
+    }
+}
+
+impl fmt::Display for Uuid {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::LowerHex::fmt(self, f)
+    }
+}
+
+impl fmt::Display for Variant {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            Variant::NCS => write!(f, "NCS"),
+            Variant::RFC4122 => write!(f, "RFC4122"),
+            Variant::Microsoft => write!(f, "Microsoft"),
+            Variant::Future => write!(f, "Future"),
+        }
+    }
+}
+
+impl fmt::LowerHex for Uuid {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::LowerHex::fmt(&self.to_hyphenated_ref(), f)
+    }
+}
+
+impl fmt::UpperHex for Uuid {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::UpperHex::fmt(&self.to_hyphenated_ref(), f)
+    }
+}
+
+impl str::FromStr for Uuid {
+    type Err = Error;
+
+    fn from_str(uuid_str: &str) -> Result<Self, Self::Err> {
+        Uuid::parse_str(uuid_str)
+    }
+}
+
+impl Default for Uuid {
+    #[inline]
+    fn default() -> Self {
+        Uuid::nil()
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    extern crate std;
-
     use super::test_util;
     use crate::prelude::*;
-    pub use std::prelude::v1::*;
+
+    macro_rules! check {
+        ($buf:ident, $format:expr, $target:expr, $len:expr, $cond:expr) => {
+            $buf.clear();
+            write!($buf, $format, $target).unwrap();
+            assert!($buf.len() == $len);
+            assert!($buf.chars().all($cond), "{}", $buf);
+        };
+    }
+
+    #[test]
+    fn test_uuid_compare() {
+        let uuid1 = test_util::new();
+        let uuid2 = test_util::new2();
+
+        assert_eq!(uuid1, uuid1);
+        assert_eq!(uuid2, uuid2);
+
+        assert_ne!(uuid1, uuid2);
+        assert_ne!(uuid2, uuid1);
+    }
+
+    #[test]
+    fn test_uuid_default() {
+        let default_uuid = Uuid::default();
+        let nil_uuid = Uuid::nil();
+
+        assert_eq!(default_uuid, nil_uuid);
+    }
+
+    #[test]
+    fn test_uuid_display() {
+        use super::fmt::Write;
+
+        let uuid = test_util::new();
+        let s = uuid.to_string();
+        let mut buffer = String::new();
+
+        assert_eq!(s, uuid.to_hyphenated().to_string());
+
+        check!(buffer, "{}", uuid, 36, |c| c.is_lowercase()
+            || c.is_digit(10)
+            || c == '-');
+    }
+
+    #[test]
+    fn test_uuid_lowerhex() {
+        use super::fmt::Write;
+
+        let mut buffer = String::new();
+        let uuid = test_util::new();
+
+        check!(buffer, "{:x}", uuid, 36, |c| c.is_lowercase()
+            || c.is_digit(10)
+            || c == '-');
+    }
+
+    // noinspection RsAssertEqual
+    #[test]
+    fn test_uuid_operator_eq() {
+        let uuid1 = test_util::new();
+        let uuid1_dup = uuid1.clone();
+        let uuid2 = test_util::new2();
+
+        assert!(uuid1 == uuid1);
+        assert!(uuid1 == uuid1_dup);
+        assert!(uuid1_dup == uuid1);
+
+        assert!(uuid1 != uuid2);
+        assert!(uuid2 != uuid1);
+        assert!(uuid1_dup != uuid2);
+        assert!(uuid2 != uuid1_dup);
+    }
+
+    #[test]
+    fn test_uuid_to_string() {
+        use super::fmt::Write;
+
+        let uuid = test_util::new();
+        let s = uuid.to_string();
+        let mut buffer = String::new();
+
+        assert_eq!(s.len(), 36);
+
+        check!(buffer, "{}", s, 36, |c| c.is_lowercase()
+            || c.is_digit(10)
+            || c == '-');
+    }
+
+    #[test]
+    fn test_uuid_upperhex() {
+        use super::fmt::Write;
+
+        let mut buffer = String::new();
+        let uuid = test_util::new();
+
+        check!(buffer, "{:X}", uuid, 36, |c| c.is_uppercase()
+            || c.is_digit(10)
+            || c == '-');
+    }
 
     #[test]
     fn test_nil() {
@@ -1046,231 +738,6 @@ mod tests {
         assert_eq!(uuid4.get_variant().unwrap(), Variant::Microsoft);
         assert_eq!(uuid5.get_variant().unwrap(), Variant::Microsoft);
         assert_eq!(uuid6.get_variant().unwrap(), Variant::NCS);
-    }
-
-    #[test]
-    fn test_parse_uuid_v4() {
-        use crate::adapter;
-        use crate::parser;
-
-        const EXPECTED_UUID_LENGTHS: parser::ExpectedLength =
-            parser::ExpectedLength::Any(&[
-                adapter::Hyphenated::LENGTH,
-                adapter::Simple::LENGTH,
-            ]);
-
-        const EXPECTED_GROUP_COUNTS: parser::ExpectedLength =
-            parser::ExpectedLength::Any(&[1, 5]);
-
-        const EXPECTED_CHARS: &'static str = "0123456789abcdefABCDEF-";
-
-        // Invalid
-        assert_eq!(
-            Uuid::parse_str(""),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 0,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("!"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 1
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF-329BF39FA1E45"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 37,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa-BBF-329BF39FA1E4"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 35
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa-BGBF-329BF39FA1E4"),
-            Err(parser::ParseError::InvalidCharacter {
-                expected: EXPECTED_CHARS,
-                found: 'G',
-                index: 20,
-                urn: parser::UrnPrefix::Optional,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2F4faaFB6BFF329BF39FA1E4"),
-            Err(parser::ParseError::InvalidGroupCount {
-                expected: EXPECTED_GROUP_COUNTS,
-                found: 2
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faaFB6BFF329BF39FA1E4"),
-            Err(parser::ParseError::InvalidGroupCount {
-                expected: EXPECTED_GROUP_COUNTS,
-                found: 3,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BFF329BF39FA1E4"),
-            Err(parser::ParseError::InvalidGroupCount {
-                expected: EXPECTED_GROUP_COUNTS,
-                found: 4,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 18,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faaXB6BFF329BF39FA1E4"),
-            Err(parser::ParseError::InvalidCharacter {
-                expected: EXPECTED_CHARS,
-                found: 'X',
-                index: 18,
-                urn: parser::UrnPrefix::Optional,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB-24fa-eB6BFF32-BF39FA1E4"),
-            Err(parser::ParseError::InvalidGroupLength {
-                expected: parser::ExpectedLength::Exact(4),
-                found: 3,
-                group: 1,
-            })
-        );
-        // (group, found, expecting)
-        //
-        assert_eq!(
-            Uuid::parse_str("01020304-1112-2122-3132-41424344"),
-            Err(parser::ParseError::InvalidGroupLength {
-                expected: parser::ExpectedLength::Exact(12),
-                found: 8,
-                group: 4,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 31,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c88"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 33,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0cg8"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 33,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("67e5504410b1426%9247bb680e5fe0c8"),
-            Err(parser::ParseError::InvalidCharacter {
-                expected: EXPECTED_CHARS,
-                found: '%',
-                index: 15,
-                urn: parser::UrnPrefix::Optional,
-            })
-        );
-
-        assert_eq!(
-            Uuid::parse_str("231231212212423424324323477343246663"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 36,
-            })
-        );
-
-        // Valid
-        assert!(Uuid::parse_str("00000000000000000000000000000000").is_ok());
-        assert!(Uuid::parse_str("67e55044-10b1-426f-9247-bb680e5fe0c8").is_ok());
-        assert!(Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4").is_ok());
-        assert!(Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c8").is_ok());
-        assert!(Uuid::parse_str("01020304-1112-2122-3132-414243444546").is_ok());
-        assert!(Uuid::parse_str(
-            "urn:uuid:67e55044-10b1-426f-9247-bb680e5fe0c8"
-        )
-        .is_ok());
-
-        // Nil
-        let nil = Uuid::nil();
-        assert_eq!(
-            Uuid::parse_str("00000000000000000000000000000000").unwrap(),
-            nil
-        );
-        assert_eq!(
-            Uuid::parse_str("00000000-0000-0000-0000-000000000000").unwrap(),
-            nil
-        );
-
-        // Round-trip
-        let uuid_orig = test_util::new();
-        let orig_str = uuid_orig.to_string();
-        let uuid_out = Uuid::parse_str(&orig_str).unwrap();
-        assert_eq!(uuid_orig, uuid_out);
-
-        // Test error reporting
-        assert_eq!(
-            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c"),
-            Err(parser::ParseError::InvalidLength {
-                expected: EXPECTED_UUID_LENGTHS,
-                found: 31,
-            })
-        );
-        assert_eq!(
-            Uuid::parse_str("67e550X410b1426f9247bb680e5fe0cd"),
-            Err(parser::ParseError::InvalidCharacter {
-                expected: EXPECTED_CHARS,
-                found: 'X',
-                index: 6,
-                urn: parser::UrnPrefix::Optional,
-            })
-        );
-        assert_eq!(
-            Uuid::parse_str("67e550-4105b1426f9247bb680e5fe0c"),
-            Err(parser::ParseError::InvalidGroupLength {
-                expected: parser::ExpectedLength::Exact(8),
-                found: 6,
-                group: 0,
-            })
-        );
-        assert_eq!(
-            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF1-02BF39FA1E4"),
-            Err(parser::ParseError::InvalidGroupLength {
-                expected: parser::ExpectedLength::Exact(4),
-                found: 5,
-                group: 3,
-            })
-        );
     }
 
     #[test]
