+++ src/parser/mod.rs
@@ -13,109 +13,13 @@
 //!
 //! [`Uuid`]: ../struct.Uuid.html
 
-mod core_support;
-#[cfg(feature = "std")]
-mod std_support;
+pub(crate) mod error;
+pub(crate) use self::error::Error;
 
-/// The expected length.
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub enum ExpectedLength {
-    /// Expected any one of the given values.
-    Any(&'static [usize]),
-    /// Expected the given value.
-    Exact(usize),
-    /// Expected any values in the given range.
-    Range {
-        /// The minimum expected value.
-        min: usize,
-        /// The maximum expected value.
-        max: usize,
-    },
-}
-
-/// Urn prefix value.
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub enum UrnPrefix {
-    /// No `urn:uuid:` prefix should be provided.
-    None,
-    /// The `urn:uuid:` prefix should optionally provided.
-    Optional,
-    /// The `urn:uuid:` prefix is required.
-    Required,
-}
-
-/// An error that can occur while parsing a [`Uuid`] string.
-///
-/// [`Uuid`]: ../struct.Uuid.html
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub enum ParseError {
-    /// Invalid character in the [`Uuid`] string.
-    ///
-    /// [`Uuid`]: ../struct.Uuid.html
-    InvalidCharacter {
-        /// The expected characters.
-        expected: &'static str,
-        /// The invalid character found.
-        found: char,
-        /// The invalid character position.
-        index: usize,
-        /// Indicates the [`Uuid`] starts with `urn:uuid:`.
-        ///
-        /// This is a special case for [`Urn`] adapter parsing.
-        ///
-        /// [`Uuid`]: ../Uuid.html
-        urn: UrnPrefix,
-    },
-    /// Invalid number of segments in the [`Uuid`] string.
-    ///
-    /// [`Uuid`]: ../struct.Uuid.html
-    InvalidGroupCount {
-        /// The expected number of segments.
-        // TODO: explain multiple segment count.
-        // BODY: Parsers can expect a range of Uuid segment count.
-        //       This needs to be expanded on.
-        expected: ExpectedLength,
-        /// The number of segments found.
-        found: usize,
-    },
-    /// Invalid length of a segment in a [`Uuid`] string.
-    ///
-    /// [`Uuid`]: ../struct.Uuid.html
-    InvalidGroupLength {
-        /// The expected length of the segment.
-        expected: ExpectedLength,
-        /// The length of segment found.
-        found: usize,
-        /// The segment with invalid length.
-        group: usize,
-    },
-    /// Invalid length of the [`Uuid`] string.
-    ///
-    /// [`Uuid`]: ../struct.Uuid.html
-    InvalidLength {
-        /// The expected length(s).
-        // TODO: explain multiple lengths.
-        // BODY: Parsers can expect a range of Uuid lenghts.
-        //       This needs to be expanded on.
-        expected: ExpectedLength,
-        /// The invalid length found.
-        found: usize,
-    },
-}
-
-impl ParseError {
-    fn _description(&self) -> &str {
-        match *self {
-            ParseError::InvalidCharacter { .. } => "invalid character",
-            ParseError::InvalidGroupCount { .. } => "invalid number of groups",
-            ParseError::InvalidGroupLength { .. } => "invalid group length",
-            ParseError::InvalidLength { .. } => "invalid length",
-        }
-    }
-}
+use crate::{adapter, Uuid};
 
 /// Check if the length matches any of the given criteria lengths.
-pub(crate) fn len_matches_any(len: usize, crits: &[usize]) -> bool {
+fn len_matches_any(len: usize, crits: &[usize]) -> bool {
     for crit in crits {
         if len == *crit {
             return true;
@@ -128,7 +32,7 @@ pub(crate) fn len_matches_any(len: usize, crits: &[usize]) -> bool {
 /// Check if the length matches any criteria lengths in the given range
 /// (inclusive).
 #[allow(dead_code)]
-pub(crate) fn len_matches_range(len: usize, min: usize, max: usize) -> bool {
+fn len_matches_range(len: usize, min: usize, max: usize) -> bool {
     for crit in min..(max + 1) {
         if len == crit {
             return true;
@@ -139,7 +43,388 @@ pub(crate) fn len_matches_range(len: usize, min: usize, max: usize) -> bool {
 }
 
 // Accumulated length of each hyphenated group in hex digits.
-pub(crate) const ACC_GROUP_LENS: [usize; 5] = [8, 12, 16, 20, 32];
+const ACC_GROUP_LENS: [usize; 5] = [8, 12, 16, 20, 32];
 
 // Length of each hyphenated group in hex digits.
-pub(crate) const GROUP_LENS: [usize; 5] = [8, 4, 4, 4, 12];
+const GROUP_LENS: [usize; 5] = [8, 4, 4, 4, 12];
+
+impl Uuid {
+    /// Parses a `Uuid` from a string of hexadecimal digits with optional
+    /// hyphens.
+    ///
+    /// Any of the formats generated by this module (simple, hyphenated, urn)
+    /// are supported by this parsing function.
+    pub fn parse_str(mut input: &str) -> Result<Uuid, crate::Error> {
+        // Ensure length is valid for any of the supported formats
+        let len = input.len();
+
+        if len == adapter::Urn::LENGTH && input.starts_with("urn:uuid:") {
+            input = &input[9..];
+        } else if !len_matches_any(
+            len,
+            &[adapter::Hyphenated::LENGTH, adapter::Simple::LENGTH],
+        ) {
+            Err(Error::InvalidLength {
+                expected: error::ExpectedLength::Any(&[
+                    adapter::Hyphenated::LENGTH,
+                    adapter::Simple::LENGTH,
+                ]),
+                found: len,
+            })?;
+        }
+
+        // `digit` counts only hexadecimal digits, `i_char` counts all chars.
+        let mut digit = 0;
+        let mut group = 0;
+        let mut acc = 0;
+        let mut buffer = [0u8; 16];
+
+        for (i_char, chr) in input.bytes().enumerate() {
+            if digit as usize >= adapter::Simple::LENGTH && group != 4 {
+                if group == 0 {
+                    Err(Error::InvalidLength {
+                        expected: error::ExpectedLength::Any(&[
+                            adapter::Hyphenated::LENGTH,
+                            adapter::Simple::LENGTH,
+                        ]),
+                        found: len,
+                    })?;
+                }
+
+                Err(Error::InvalidGroupCount {
+                    expected: error::ExpectedLength::Any(&[1, 5]),
+                    found: group + 1,
+                })?;
+            }
+
+            if digit % 2 == 0 {
+                // First digit of the byte.
+                match chr {
+                    // Calulate upper half.
+                    b'0'..=b'9' => acc = chr - b'0',
+                    b'a'..=b'f' => acc = chr - b'a' + 10,
+                    b'A'..=b'F' => acc = chr - b'A' + 10,
+                    // Found a group delimiter
+                    b'-' => {
+                        // TODO: remove the u8 cast
+                        // BODY: this only needed until we switch to
+                        //       ParseError
+                        if ACC_GROUP_LENS[group] as u8 != digit {
+                            // Calculate how many digits this group consists of
+                            // in the input.
+                            let found = if group > 0 {
+                                // TODO: remove the u8 cast
+                                // BODY: this only needed until we switch to
+                                //       ParseError
+                                digit - ACC_GROUP_LENS[group - 1] as u8
+                            } else {
+                                digit
+                            };
+
+                            Err(
+                                Error::InvalidGroupLength {
+                                    expected: error::ExpectedLength::Exact(
+                                        GROUP_LENS[group],
+                                    ),
+                                    found: found as usize,
+                                    group,
+                                },
+                            )?;
+                        }
+                        // Next group, decrement digit, it is incremented again
+                        // at the bottom.
+                        group += 1;
+                        digit -= 1;
+                    }
+                    _ => {
+                        Err(Error::InvalidCharacter {
+                            expected: "0123456789abcdefABCDEF-",
+                            found: input[i_char..].chars().next().unwrap(),
+                            index: i_char,
+                            urn: error::UrnPrefix::Optional,
+                        })?;
+                    }
+                }
+            } else {
+                // Second digit of the byte, shift the upper half.
+                acc *= 16;
+                match chr {
+                    b'0'..=b'9' => acc += chr - b'0',
+                    b'a'..=b'f' => acc += chr - b'a' + 10,
+                    b'A'..=b'F' => acc += chr - b'A' + 10,
+                    b'-' => {
+                        // The byte isn't complete yet.
+                        let found = if group > 0 {
+                            // TODO: remove the u8 cast
+                            // BODY: this only needed until we switch to
+                            //       ParseError
+                            digit - ACC_GROUP_LENS[group - 1] as u8
+                        } else {
+                            digit
+                        };
+
+                        Err(Error::InvalidGroupLength {
+                            expected: error::ExpectedLength::Exact(
+                                GROUP_LENS[group],
+                            ),
+                            found: found as usize,
+                            group,
+                        })?;
+                    }
+                    _ => {
+                        Err(Error::InvalidCharacter {
+                            expected: "0123456789abcdefABCDEF-",
+                            found: input[i_char..].chars().next().unwrap(),
+                            index: i_char,
+                            urn: error::UrnPrefix::Optional,
+                        })?;
+                    }
+                }
+                buffer[(digit / 2) as usize] = acc;
+            }
+            digit += 1;
+        }
+
+        // Now check the last group.
+        // TODO: remove the u8 cast
+        // BODY: this only needed until we switch to
+        //       ParseError
+        if ACC_GROUP_LENS[4] as u8 != digit {
+            Err(Error::InvalidGroupLength {
+                expected: error::ExpectedLength::Exact(GROUP_LENS[4]),
+                found: (digit as usize - ACC_GROUP_LENS[3]),
+                group,
+            })?;
+        }
+
+        Ok(Uuid::from_bytes(buffer))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{adapter, test_util};
+
+    #[test]
+    fn test_parse_uuid_v4() {
+        const EXPECTED_UUID_LENGTHS: error::ExpectedLength =
+            error::ExpectedLength::Any(&[
+                adapter::Hyphenated::LENGTH,
+                adapter::Simple::LENGTH,
+            ]);
+
+        const EXPECTED_GROUP_COUNTS: error::ExpectedLength =
+            error::ExpectedLength::Any(&[1, 5]);
+
+        const EXPECTED_CHARS: &'static str = "0123456789abcdefABCDEF-";
+
+        // Invalid
+        assert_eq!(
+            Uuid::parse_str("").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 0,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("!").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 1
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF-329BF39FA1E45").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 37,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa-BBF-329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 35
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa-BGBF-329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidCharacter {
+                expected: EXPECTED_CHARS,
+                found: 'G',
+                index: 20,
+                urn: error::UrnPrefix::Optional,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2F4faaFB6BFF329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupCount {
+                expected: EXPECTED_GROUP_COUNTS,
+                found: 2
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faaFB6BFF329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupCount {
+                expected: EXPECTED_GROUP_COUNTS,
+                found: 3,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BFF329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupCount {
+                expected: EXPECTED_GROUP_COUNTS,
+                found: 4,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 18,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faaXB6BFF329BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidCharacter {
+                expected: EXPECTED_CHARS,
+                found: 'X',
+                index: 18,
+                urn: error::UrnPrefix::Optional,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB-24fa-eB6BFF32-BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupLength {
+                expected: error::ExpectedLength::Exact(4),
+                found: 3,
+                group: 1,
+            })
+        );
+        // (group, found, expecting)
+        //
+        assert_eq!(
+            Uuid::parse_str("01020304-1112-2122-3132-41424344").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupLength {
+                expected: error::ExpectedLength::Exact(12),
+                found: 8,
+                group: 4,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 31,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c88").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 33,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0cg8").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 33,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("67e5504410b1426%9247bb680e5fe0c8").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidCharacter {
+                expected: EXPECTED_CHARS,
+                found: '%',
+                index: 15,
+                urn: error::UrnPrefix::Optional,
+            })
+        );
+
+        assert_eq!(
+            Uuid::parse_str("231231212212423424324323477343246663").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 36,
+            })
+        );
+
+        // Valid
+        assert!(Uuid::parse_str("00000000000000000000000000000000").is_ok());
+        assert!(Uuid::parse_str("67e55044-10b1-426f-9247-bb680e5fe0c8").is_ok());
+        assert!(Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4").is_ok());
+        assert!(Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c8").is_ok());
+        assert!(Uuid::parse_str("01020304-1112-2122-3132-414243444546").is_ok());
+        assert!(Uuid::parse_str(
+            "urn:uuid:67e55044-10b1-426f-9247-bb680e5fe0c8"
+        )
+        .is_ok());
+
+        // Nil
+        let nil = Uuid::nil();
+        assert_eq!(
+            Uuid::parse_str("00000000000000000000000000000000").unwrap(),
+            nil
+        );
+        assert_eq!(
+            Uuid::parse_str("00000000-0000-0000-0000-000000000000").unwrap(),
+            nil
+        );
+
+        // Round-trip
+        let uuid_orig = test_util::new();
+        let orig_str = uuid_orig.to_string();
+        let uuid_out = Uuid::parse_str(&orig_str).unwrap();
+        assert_eq!(uuid_orig, uuid_out);
+
+        // Test error reporting
+        assert_eq!(
+            Uuid::parse_str("67e5504410b1426f9247bb680e5fe0c").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidLength {
+                expected: EXPECTED_UUID_LENGTHS,
+                found: 31,
+            })
+        );
+        assert_eq!(
+            Uuid::parse_str("67e550X410b1426f9247bb680e5fe0cd").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidCharacter {
+                expected: EXPECTED_CHARS,
+                found: 'X',
+                index: 6,
+                urn: error::UrnPrefix::Optional,
+            })
+        );
+        assert_eq!(
+            Uuid::parse_str("67e550-4105b1426f9247bb680e5fe0c").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupLength {
+                expected: error::ExpectedLength::Exact(8),
+                found: 6,
+                group: 0,
+            })
+        );
+        assert_eq!(
+            Uuid::parse_str("F9168C5E-CEB2-4faa-B6BF1-02BF39FA1E4").map_err(crate::Error::expect_parser),
+            Err(Error::InvalidGroupLength {
+                expected: error::ExpectedLength::Exact(4),
+                found: 5,
+                group: 3,
+            })
+        );
+    }
+}
