+++ src/winapi_support.rs
@@ -6,7 +6,7 @@ impl Uuid {
     /// Attempts to create a [`Uuid`] from a little endian winapi `GUID`
     ///
     /// [`Uuid`]: ../struct.Uuid.html
-    pub fn from_guid(guid: guiddef::GUID) -> Result<Uuid, crate::BytesError> {
+    pub fn from_guid(guid: guiddef::GUID) -> Result<Uuid, crate::Error> {
         Uuid::from_fields_le(
             guid.Data1 as u32,
             guid.Data2 as u16,
